requirements.txt
PyQt5~=5.15.11
winrt
Pillow
pyperclip